pyproject.toml
[project]
name = "cartoon-sun-position"
version = "0.1.0"
description = "Generate a cartoonish sun position image for current time"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
generate = "cartoon_sun_position.__main__:main"
